Change the ID scheme used for Elasticsearch.
Swift has a natural ID scheme for Elasticsearch documents:
"account/container/object" are unique identifiers of a given object.
Unfortunately, this scheme fails when object (or container or account)
names are very long. Elasticsearch imposes a 512 character limit and
we need to adjust our ID scheme to be less than that.