Handle unicode characters in object names.
There is a subtle bug where the Swift object names are UTF-8 strings,
whereas account and container values we load from the JSON file are
unicode. When we join these, and if there is a unicode character in
the Swift object name, we get back a unicode string. Unfortunately,
hashlib expects a "str" type, not unicode.